requirements.txt
stable-baselines3
gym[box2d]
ipython
ipykernel
SoundFile
opencv-python
pillow
moviepy
gym-retro
pygame
sounddevice
mpi4py
neptune-client
psutil
pyglet
numpy
